Restoring Github Primitives
# Changes in Cleaning Files:
Early September 2014: I began making changes to the cleaning files to
(a) incorporate different dates with the payments (i.e. Challan, Entry,
Attest) and (b) incorporate new 2014 data that wasn’t in previous
versions of the data.  I started this work sometime around 9/16 and it
turned out to be a lot more involved than I was hoping.  When I began
making substantial changes, I pushed my worked to Github and reverted
all the change on Dropbox so I could work with a local copy without
corrupting the previous work.

The main changes were to the following files:

- Dropbox\BD Taxation\Code\pre analysis\Payment and Registration\...
- merge\_pay\_reg_census.do
- Version 17 (which is really Version 0) on DB is the primitive for
Github
- generate\_registration\_file.do
- Verson 3 (which is really Version 0) on DB is the primitive for Github
- merge\_id\_to\_uncollapsed_payments.do
- Version 2 (which is restored from Version 0) on DB is the primitive
for Github
- generate\_payment\_by\_bin.do
- Versions 36 (which is restored from Version 0 plus some whitespace
differences) on DB is the primitive for Github